[OpenCL] Fix checking of vector type casting
Currently clang allows the following code

int a;
int b = (const int) a;
However it does not the following code

int4 a;
int4 b = (const int4) a;
This is because Clang compares the qualified types instead of unqualified types for vector type casting, which causes the inconsistency.

This patch fixes that.

Differential Revision: https://reviews.llvm.org/D38463


git-svn-id: http://llvm.org/svn/llvm-project/cfe/trunk@314802 91177308-0d34-0410-b5e6-96231b3b80d8