[Driver] Fix codeview-column-info on macOS
macOS home directory paths begin with /Users, and clang-cl interprets
the /U portion as a macro undefine rather than a path, causing test
failures on macOS. Use a -- to explicitly treat the input file as a path
and fix the test.

This effectively reverts r326168 and adds an alternative fix.

git-svn-id: http://llvm.org/svn/llvm-project/cfe/trunk@326171 91177308-0d34-0410-b5e6-96231b3b80d8