Remove failing test.
Removing on behalf of Jorge Moya.  This test is broken on
Windows due to it actually being able to resolve the path.  There
is an actual Windows-specific bug somewhere, but we already have
sufficient test coverage of this with a different test, so removing
this was the approach suggested by Jorge.

git-svn-id: http://llvm.org/svn/llvm-project/cfe/trunk@342693 91177308-0d34-0410-b5e6-96231b3b80d8