[analyzer][NFC] Refactor PlistDiagnostics to use a class instead of passing 9 parameters around
This has been a long time coming. Note the usage of AnalyzerOptions: I'll need
it for D52742, and added it in rC343620. The main motivation for this was that
I'll need to add yet another parameter to every single function, and some
functions would reach their 10th parameter with that change.


git-svn-id: http://llvm.org/svn/llvm-project/cfe/trunk@345531 91177308-0d34-0410-b5e6-96231b3b80d8