Speculatively attempt to fix a failing testbot.
A testbot (	http://green.lab.llvm.org/green/job/clang-stage1-cmake-RA-incremental/54690/) was failing with a complaint about an obsolete option that wasn't present in the command line in the first place. This replaces my guess at the "obsolete option" with a different spelling that will hopefully be more acceptable to this bot without breaking other bots.

git-svn-id: http://llvm.org/svn/llvm-project/cfe/trunk@345635 91177308-0d34-0410-b5e6-96231b3b80d8